CRUD completo en Comment y Proceeding
Se implementó todas las funciones que comforman el acrónimo CRUD en las clases Comment y Proceeding

+++ Lawdingo_G4/src/main/java/pe/edu/upc/aaw/lawdingo_g4/controllers/RoleController.java
@@ -34,7 +34,7 @@ public class RoleController {
         }).collect(Collectors.toList());
     }
     @GetMapping("/{id}")
-    public RoleDTO listId(@PathVariable("id")Long id){
+    public RoleDTO listId(@PathVariable("id")Integer id){
         ModelMapper m = new ModelMapper();
         RoleDTO dto = m.map(rService.listId(id), RoleDTO.class);
         return dto;
