+++ Lawdingo_G4/src/main/java/pe/edu/upc/aaw/lawdingo_g4/controllers/UserController.java
@@ -39,7 +39,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") Long id){
+    public void delete(@PathVariable("id") Integer id){
         uS.delete(id);
     }
