+++ Lawdingo_G4/src/main/java/pe/edu/upc/aaw/lawdingo_g4/serviceimplements/RoleServiceImplement.java
@@ -25,7 +25,7 @@ public class RoleServiceImplement implements IRoleService {
     }
 
     @Override
-    public Role listId(long id) {
+    public Role listId(int id) {
         return rR.findById(id).orElse(new Role());
     }
 }
