+++ Lawdingo_G4/src/main/java/pe/edu/upc/aaw/lawdingo_g4/serviceimplements/UserServiceImplement.java
@@ -31,7 +31,7 @@ public class UserServiceImplement implements IUserService {
 
 
     @Override
-    public void  delete(Long idUser) {
+    public void  delete(int idUser) {
         uR.deleteById(idUser);
     }
