+++ Lawdingo_G4/src/main/java/pe/edu/upc/aaw/lawdingo_g4/serviceinterfaces/IRoleService.java
@@ -8,6 +8,6 @@ public interface IRoleService {
     public void insert(Role role);
 
     List<Role> list();
-    public Role listId(long id);
+    public Role listId(int id);
 
 }
