+++ Lawdingo_G4/src/main/java/pe/edu/upc/aaw/lawdingo_g4/serviceinterfaces/IUserService.java
@@ -9,6 +9,6 @@ public interface IUserService {
     public List <Users> getUsersWhoseNameStartsWith(String startsWith);
     public void insert (Users users);
     public List <Users> list(String name);
-    public void  delete (Long idUser);
+    public void  delete (int idUser);
 
 }
